+++ src/views/Typography.jsx
@@ -1,12 +1,15 @@
 import React from "react";
+
+// reactstrap components
 import { Card, CardHeader, CardBody, Row, Col } from "reactstrap";
 
-import { PanelHeader } from "components";
+// core components
+import PanelHeader from "components/PanelHeader/PanelHeader.jsx";
 
 class Typography extends React.Component {
   render() {
     return (
-      <div>
+      <>
         <PanelHeader size="sm" />
         <div className="content">
           <Row>
@@ -51,12 +54,12 @@ class Typography extends React.Component {
                   </div>
                   <div className="typography-line">
                     <p>
-                      <span>Paragraph</span>
-                      I will be the leader of a company that ends up being worth
-                      billions of dollars, because I got the answers. I
-                      understand culture. I am the nucleus. I think that’s a
-                      responsibility that I have, to push possibilities, to show
-                      people, this is the level that things could be at.
+                      <span>Paragraph</span>I will be the leader of a company
+                      that ends up being worth billions of dollars, because I
+                      got the answers. I understand culture. I am the nucleus. I
+                      think that’s a responsibility that I have, to push
+                      possibilities, to show people, this is the level that
+                      things could be at.
                     </p>
                   </div>
                   <div className="typography-line">
@@ -129,7 +132,7 @@ class Typography extends React.Component {
             </Col>
           </Row>
         </div>
-      </div>
+      </>
     );
   }
 }