+++ src/views/UserPage.jsx
@@ -1,155 +1,164 @@
 import React from "react";
-import { Card, CardHeader, CardBody, Row, Col } from "reactstrap";
 
-import { PanelHeader, FormInputs, CardAuthor, CardSocials } from "components";
+// reactstrap components
+import {
+  Button,
+  Card,
+  CardHeader,
+  CardBody,
+  FormGroup,
+  Form,
+  Input,
+  Row,
+  Col
+} from "reactstrap";
 
-import userBackground from "assets/img/bg5.jpg";
-import userAvatar from "assets/img/mike.jpg";
+// core components
+import PanelHeader from "components/PanelHeader/PanelHeader.jsx";
 
 class User extends React.Component {
   render() {
     return (
-      <div>
+      <>
         <PanelHeader size="sm" />
         <div className="content">
           <Row>
-            <Col md={8} xs={12}>
+            <Col md="8">
               <Card>
                 <CardHeader>
                   <h5 className="title">Edit Profile</h5>
                 </CardHeader>
                 <CardBody>
-                  <form>
-                    <FormInputs
-                      ncols={[
-                        "col-md-5 pr-1",
-                        "col-md-3 px-1",
-                        "col-md-4 pl-1"
-                      ]}
-                      proprieties={[
-                        {
-                          label: "Company (disabled)",
-                          inputProps: {
-                            type: "text",
-                            disabled: true,
-                            defaultValue: "Creative Code Inc."
-                          }
-                        },
-                        {
-                          label: "Username",
-                          inputProps: {
-                            type: "text",
-                            defaultValue: "michael23"
-                          }
-                        },
-                        {
-                          label: "Email address",
-                          inputProps: {
-                            type: "email",
-                            placeholder: "Email"
-                          }
-                        }
-                      ]}
-                    />
-                    <FormInputs
-                      ncols={["col-md-6 pr-1", "col-md-6 pl-1"]}
-                      proprieties={[
-                        {
-                          label: "First Name",
-                          inputProps: {
-                            type: "text",
-                            placeholder: "First Name",
-                            defaultValue: "Mike"
-                          }
-                        },
-                        {
-                          label: "Last Name",
-                          inputProps: {
-                            type: "text",
-                            placeholder: "Last Name",
-                            defaultValue: "Andrew"
-                          }
-                        }
-                      ]}
-                    />
-                    <FormInputs
-                      ncols={["col-md-12"]}
-                      proprieties={[
-                        {
-                          label: "Address",
-                          inputProps: {
-                            type: "text",
-                            placeholder: "Home Address",
-                            defaultValue:
-                              "Bld Mihail Kogalniceanu, nr. 8 Bl 1, Sc 1, Ap 09"
-                          }
-                        }
-                      ]}
-                    />
-                    <FormInputs
-                      ncols={[
-                        "col-md-4 pr-1",
-                        "col-md-4 px-1",
-                        "col-md-4 pl-1"
-                      ]}
-                      proprieties={[
-                        {
-                          label: "City",
-                          inputProps: {
-                            type: "text",
-                            defaultValue: "Bucharest",
-                            placeholder: "City"
-                          }
-                        },
-                        {
-                          label: "Country",
-                          inputProps: {
-                            type: "text",
-                            defaultValue: "Romania",
-                            placeholder: "Country"
-                          }
-                        },
-                        {
-                          label: "Postal Code",
-                          inputProps: {
-                            type: "number",
-                            placeholder: "ZIP Code"
-                          }
-                        }
-                      ]}
-                    />
-                    <FormInputs
-                      ncols={["col-md-12"]}
-                      proprieties={[
-                        {
-                          label: "About Me",
-                          inputProps: {
-                            type: "textarea",
-                            rows: "4",
-                            cols: "80",
-                            defaultValue:
-                              "Lamborghini Mercy, Your chick she so thirsty, I'm in that two seat Lambo.",
-                            placeholder: "Here can be your description"
-                          }
-                        }
-                      ]}
-                    />
-                  </form>
+                  <Form>
+                    <Row>
+                      <Col className="pr-1" md="5">
+                        <FormGroup>
+                          <label>Company (disabled)</label>
+                          <Input
+                            defaultValue="Creative Code Inc."
+                            disabled
+                            placeholder="Company"
+                            type="text"
+                          />
+                        </FormGroup>
+                      </Col>
+                      <Col className="px-1" md="3">
+                        <FormGroup>
+                          <label>Username</label>
+                          <Input
+                            defaultValue="michael23"
+                            placeholder="Username"
+                            type="text"
+                          />
+                        </FormGroup>
+                      </Col>
+                      <Col className="pl-1" md="4">
+                        <FormGroup>
+                          <label htmlFor="exampleInputEmail1">
+                            Email address
+                          </label>
+                          <Input placeholder="Email" type="email" />
+                        </FormGroup>
+                      </Col>
+                    </Row>
+                    <Row>
+                      <Col className="pr-1" md="6">
+                        <FormGroup>
+                          <label>First Name</label>
+                          <Input
+                            defaultValue="Mike"
+                            placeholder="Company"
+                            type="text"
+                          />
+                        </FormGroup>
+                      </Col>
+                      <Col className="pl-1" md="6">
+                        <FormGroup>
+                          <label>Last Name</label>
+                          <Input
+                            defaultValue="Andrew"
+                            placeholder="Last Name"
+                            type="text"
+                          />
+                        </FormGroup>
+                      </Col>
+                    </Row>
+                    <Row>
+                      <Col md="12">
+                        <FormGroup>
+                          <label>Address</label>
+                          <Input
+                            defaultValue="Bld Mihail Kogalniceanu, nr. 8 Bl 1, Sc 1, Ap 09"
+                            placeholder="Home Address"
+                            type="text"
+                          />
+                        </FormGroup>
+                      </Col>
+                    </Row>
+                    <Row>
+                      <Col className="pr-1" md="4">
+                        <FormGroup>
+                          <label>City</label>
+                          <Input
+                            defaultValue="Mike"
+                            placeholder="City"
+                            type="text"
+                          />
+                        </FormGroup>
+                      </Col>
+                      <Col className="px-1" md="4">
+                        <FormGroup>
+                          <label>Country</label>
+                          <Input
+                            defaultValue="Andrew"
+                            placeholder="Country"
+                            type="text"
+                          />
+                        </FormGroup>
+                      </Col>
+                      <Col className="pl-1" md="4">
+                        <FormGroup>
+                          <label>Postal Code</label>
+                          <Input placeholder="ZIP Code" type="number" />
+                        </FormGroup>
+                      </Col>
+                    </Row>
+                    <Row>
+                      <Col md="12">
+                        <FormGroup>
+                          <label>About Me</label>
+                          <Input
+                            cols="80"
+                            defaultValue="Lamborghini Mercy, Your chick she so thirsty, I'm in
+                            that two seat Lambo."
+                            placeholder="Here can be your description"
+                            rows="4"
+                            type="textarea"
+                          />
+                        </FormGroup>
+                      </Col>
+                    </Row>
+                  </Form>
                 </CardBody>
               </Card>
             </Col>
-            <Col md={4} xs={12}>
+            <Col md="4">
               <Card className="card-user">
                 <div className="image">
-                  <img src={userBackground} alt="..." />
+                  <img alt="..." src={require("assets/img/bg5.jpg")} />
                 </div>
                 <CardBody>
-                  <CardAuthor
-                    avatar={userAvatar}
-                    avatarAlt="..."
-                    title="Mike Andrew"
-                    description="michael23"
-                  />
+                  <div className="author">
+                    <a href="#pablo" onClick={e => e.preventDefault()}>
+                      <img
+                        alt="..."
+                        className="avatar border-gray"
+                        src={require("assets/img/mike.jpg")}
+                      />
+                      <h5 className="title">Mike Andrew</h5>
+                    </a>
+                    <p className="description">michael24</p>
+                  </div>
                   <p className="description text-center">
                     "Lamborghini Mercy <br />
                     Your chick she so thirsty <br />
@@ -157,28 +166,40 @@ class User extends React.Component {
                   </p>
                 </CardBody>
                 <hr />
-                <CardSocials
-                  size="lg"
-                  socials={[
-                    {
-                      icon: "fab fa-facebook-f",
-                      href: "https://www.facebook.com/"
-                    },
-                    {
-                      icon: "fab fa-twitter",
-                      href: "https://www.facebook.com/"
-                    },
-                    {
-                      icon: "fab fa-google-plus-g",
-                      href: "https://plus.google.com/discover"
-                    }
-                  ]}
-                />
+                <div className="button-container">
+                  <Button
+                    className="btn-neutral btn-icon btn-round"
+                    color="default"
+                    href="#pablo"
+                    onClick={e => e.preventDefault()}
+                    size="lg"
+                  >
+                    <i className="fab fa-facebook-f" />
+                  </Button>
+                  <Button
+                    className="btn-neutral btn-icon btn-round"
+                    color="default"
+                    href="#pablo"
+                    onClick={e => e.preventDefault()}
+                    size="lg"
+                  >
+                    <i className="fab fa-twitter" />
+                  </Button>
+                  <Button
+                    className="btn-neutral btn-icon btn-round"
+                    color="default"
+                    href="#pablo"
+                    onClick={e => e.preventDefault()}
+                    size="lg"
+                  >
+                    <i className="fab fa-google-plus-g" />
+                  </Button>
+                </div>
               </Card>
             </Col>
           </Row>
         </div>
-      </div>
+      </>
     );
   }
 }